requirements.txt
SQLAlchemy
numpy
pandas
python-dotenv
PyMySQL
pydantic
fastapi
uvicorn[standard]
mysql